refactor(cqrs): added new modules
